requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pgvector==0.2.4
google-generativeai>=0.5.0==0.8.3
requests==2.31.0
beautifulsoup4==4.12.3
pypdf==4.0.1
openpyxl==3.1.2
olefile==0.47
streamlit==1.31.0
pandas==2.2.0
langchain-text-splitters
PyMuPDF
python-docx
openai
anthropic
python-multipart
python-jose[cryptography]
bcrypt
asgiref
celery[redis]

# Enterprise Refactoring Dependencies
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
pydantic-settings==2.1.0
greenlet==3.0.3